impr: Replace Integer to Long type in Repositories

+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/ActiviteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Activite;
 
 
-public interface ActiviteRepository extends CrudRepository<Activite, Integer> {
+public interface ActiviteRepository extends CrudRepository<Activite, Long> {
     
 }
