+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/FichierRepository.java
@@ -3,7 +3,7 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Fichier;
 
-public interface FichierRepository extends CrudRepository<Fichier, Integer> {
+public interface FichierRepository extends CrudRepository<Fichier, Long> {
 
     
 }
