+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/LogRepository.java
@@ -3,7 +3,7 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Log;
 
-public interface LogRepository extends CrudRepository<Log, Integer> {
+public interface LogRepository extends CrudRepository<Log, Long> {
 
     
 }
