+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/NotificationRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.Notification;
 
-public interface NotificationRepository extends CrudRepository<Notification, Integer> {
+public interface NotificationRepository extends CrudRepository<Notification, Long> {
     
 }
