+++ java-backend/src/main/java/com/ajmservices/javabackend/repository/RelationUtilisateurRepository.java
@@ -3,6 +3,6 @@ package com.ajmservices.javabackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.ajmservices.javabackend.model.RelationUtilisateur;
 
-public interface RelationUtilisateurRepository extends CrudRepository<RelationUtilisateur, Integer> {
+public interface RelationUtilisateurRepository extends CrudRepository<RelationUtilisateur, Long> {
 
 }
